tiker/requirements.txt
altair==4.2.0
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
attrs==22.2.0
beautifulsoup4==4.11.1
blinker==1.5
bs4==0.0.1
cachetools==5.2.1
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
clean-text==0.6.0
cleantext==1.1.4
click==8.1.3
cmdstanpy==1.0.8
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.6
convertdate==2.4.0
crypto==1.4.1
cryptocompare==0.7.6
cryptography==39.0.0
cssselect==1.2.0
cycler==0.11.0
dateparser==1.1.6
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
django-localflavor==3.1
django-phone-field==1.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
emoji==1.7.0
entrypoints==0.4
ephem==4.1.4
fake-useragent==1.1.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
fonttools==4.38.0
frozendict==2.3.4
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.30
GoogleNews==1.6.6
h11==0.14.0
hijri-converter==2.2.4
holidays==0.18
html5lib==1.1
httpcore==0.16.3
httpx==0.23.1
idna==3.4
importlib-metadata==6.0.0
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.3
mftool==2.5
multitasking==0.0.11
Naked==0.1.32
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.1
packaging==23.0
pandas==1.5.2
pandas-datareader==0.10.0
parse==1.19.0
Pillow==9.4.0
platformdirs==2.6.0
plotly==5.11.0
prophet==1.1.1
protobuf==3.20.3
pyarrow==10.0.1
pycparser==2.21
pycryptodome==3.16.0
pydeck==0.8.0
pyee==8.2.2
Pygments==2.14.0
PyJWT==2.6.0
PyMeeus==0.5.12
Pympler==1.0.1
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-stdnum==1.18
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
QuantStats==0.0.59
regex==2022.10.31
requests==2.28.2
requests-cache==0.9.7
requests-file==1.5.1
requests-html==0.10.0
rfc3986==1.5.0
rich==13.0.1
scipy==1.10.0
seaborn==0.12.2
semver==2.13.0
setuptools-git==1.2
sgmllib3k==1.0.0
shellescape==3.8.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
streamlit==1.17.0
tabulate==0.9.0
tenacity==8.1.0
tinysegmenter==0.3
tldextract==3.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
url-normalize==1.4.3
urllib3==1.26.14
urlopen==1.0.0
vaderSentiment==3.3.2
validators==0.20.0
virtualenv==20.17.1
w3lib==2.1.1
watchdog==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
wrapt==1.14.1
yahoo-fin==0.8.9.1
yfinance==0.2.3
zipp==3.11.0
